Move benchmarks to an inner crate
That's because there are following requirements:
* Tests should work for pre-1.26 and post-1.26 Rust alike;
* Benchmarks depend on 128-bit integer support from Rand.

The issue is that Rand 0.5 supports that via Cargo feature flag, not
`rustc` version detection and `cfg`-dependent feature selection doesn't
work: <https://github.com/rust-lang/cargo/issues/2524>.

A solution is to have a crate just for benchmarking. The directory
structure looks like this because there are plans to use Criterion for
benchmarking on stable and beta Rusts.