Automatic Travis documentation build
Improve spans in usage-originated errors

Those are not emitted as `compile_error!`s so the multi-token spanning
feature is not directly available. However, it can be replicated using a
`quote_spanned!`-like macro that takes a `T: quote::ToTokens` expression
to extract span information from, then applies start span to the first
token of the output and end span to the rest tokens of the output.